requirements.txt
# Person ReID System Requirements

# Deep Learning
torch>=1.10.0
torchvision>=0.11.0
torchreid>=1.4.0

# Computer Vision
opencv-python>=4.5.0
numpy>=1.21.0
scipy>=1.7.0

# Vector Database
qdrant-client>=1.0.0

# Utilities
loguru>=0.6.0
pyyaml>=6.0
python-dotenv>=0.19.0
tqdm>=4.62.0

# Optional
matplotlib>=3.4.0
pandas>=1.3.0

